requirements.txt
langchain==0.3.25
langchain-community==0.3.25
langchain-ollama==0.3.3
chromadb==1.0.12
openai==1.12.0
pip==24.3.1

